+++ stable_baselines3/common/on_policy_algorithm.py
@@ -162,6 +162,8 @@ class OnPolicyAlgorithm(BaseAlgorithm):
 
             new_obs, rewards, dones, infos = env.step(clipped_actions)
 
+            # Give access to local variables
+            callback.update_locals(locals())
             if callback.on_step() is False:
                 return False