+++ stable_baselines3/common/callbacks.py
@@ -33,8 +33,8 @@ class BaseCallback(ABC):
         # n_envs * n times env.step() was called
         self.num_timesteps = 0  # type: int
         self.verbose = verbose
-        self.locals = None  # type: Optional[Dict[str, Any]]
-        self.globals = None  # type: Optional[Dict[str, Any]]
+        self.locals: Dict[str, Any] = {}
+        self.globals: Dict[str, Any] = {}
         self.logger = None
         # Sometimes, for event callback, it is useful
         # to have access to the parent object
@@ -103,6 +103,23 @@ class BaseCallback(ABC):
     def _on_rollout_end(self) -> None:
         pass
 
+    def update_locals(self, locals_: Dict[str, Any]) -> None:
+        """
+        Update the references to the local variables.
+
+        :param locals_: (Dict[str, Any]) the local variables during rollout collection
+        """
+        self.locals.update(locals_)
+        self.update_child_locals(locals_)
+
+    def update_child_locals(self, locals_: Dict[str, Any]) -> None:
+        """
+        Update the references to the local variables on sub callbacks.
+
+        :param locals_: (Dict[str, Any]) the local variables during rollout collection
+        """
+        pass
+
 
 class EventCallback(BaseCallback):
     """
@@ -136,6 +153,15 @@ class EventCallback(BaseCallback):
 
     def _on_step(self) -> bool:
         return True
+
+    def update_child_locals(self, locals_: Dict[str, Any]) -> None:
+        """
+        Update the references to the local variables.
+
+        :param locals_: (Dict[str, Any]) the local variables during rollout collection
+        """
+        if self.callback is not None:
+            self.callback.update_locals(locals_)
 
 
 class CallbackList(BaseCallback):
@@ -177,6 +203,15 @@ class CallbackList(BaseCallback):
     def _on_training_end(self) -> None:
         for callback in self.callbacks:
             callback.on_training_end()
+
+    def update_child_locals(self, locals_: Dict[str, Any]) -> None:
+        """
+        Update the references to the local variables.
+
+        :param locals_: (Dict[str, Any]) the local variables during rollout collection
+        """
+        for callback in self.callbacks:
+            callback.update_locals(locals_)
 
 
 class CheckpointCallback(BaseCallback):
@@ -342,6 +377,15 @@ class EvalCallback(EventCallback):
                     return self._on_event()
 
         return True
+
+    def update_child_locals(self, locals_: Dict[str, Any]) -> None:
+        """
+        Update the references to the local variables.
+
+        :param locals_: (Dict[str, Any]) the local variables during rollout collection
+        """
+        if self.callback:
+            self.callback.update_locals(locals_)
 
 
 class StopTrainingOnRewardThreshold(BaseCallback):